Non Preemptive Priority Scheduling
In the Non Preemptive Priority scheduling, The Processes are scheduled according to the priority number assigned to them. Once the process gets scheduled, it will run till the completion. Generally, the lower the priority number, the higher is the priority of the process. The people might get confused with the priority numbers, hence in the GATE, there clearly mention which one is the highest priority and which one is the lowest one.
In the Example, there are 5 Thread ( a ,  b , c , d and  f  )Their priorities ( 6 , 1 , 3 , 5 and 10 )